server/internal/registry: implement CloseNotify and Flush (for now) (#9402)
This fixes panics introduced in e90666cc70c38b53d3dfd10ffc52967d48a343f9
when Gin ungracefully assumes that the http.ResponseWriter implements
http.CloseNotifier and http.Flusher, which our new statusCodeRecorder
does not. This is a temporary fix until we can pour the rest of the Gin
out.